Use :delayed restart for neutron-dhcp-agent, instead of :immediately
We've been seeing Chef provisioning failures where an immediate
restart of neutron-dhcp-agent fails like this:

[2015-01-23T10:16:24+00:00] INFO: Processing apt_package[neutron-dhcp-agent] action install (calico::general_compute line 132)
[2015-01-23T10:16:25+00:00] INFO: Processing template[/etc/neutron/neutron.conf] action create (calico::general_compute line 137)
[2015-01-23T10:16:25+00:00] INFO: template[/etc/neutron/neutron.conf] backed up to /var/chef/backup/etc/neutron/neutron.conf.chef-20150123101625.191165
[2015-01-23T10:16:25+00:00] INFO: template[/etc/neutron/neutron.conf] updated file contents /etc/neutron/neutron.conf
[2015-01-23T10:16:25+00:00] INFO: template[/etc/neutron/neutron.conf] sending restart action to service[neutron-dhcp-agent] (immediate)
[2015-01-23T10:16:25+00:00] INFO: Processing service[neutron-dhcp-agent] action restart (calico::general_compute line 157)

================================================================================
Error executing action `restart` on resource 'service[neutron-dhcp-agent]'
================================================================================

Mixlib::ShellOut::ShellCommandFailed
------------------------------------
Expected process to exit with [0], but received '1'
---- Begin output of /sbin/start neutron-dhcp-agent ----
STDOUT:
STDERR: start: Job is already running: neutron-dhcp-agent
---- End output of /sbin/start neutron-dhcp-agent ----
Ran /sbin/start neutron-dhcp-agent returned 1

At the same time, /var/log/neutron/dhcp-agent.log shows that the agent
is rapidly failing and restarting:

2015-01-23 10:16:25.192 18068 ERROR neutron.agent.dhcp_agent [-] Error importing interface driver 'None': 'NoneType' object has no attribute 'rpartition'
2015-01-23 10:16:25.406 18094 INFO neutron.common.config [-] Logging enabled!
2015-01-23 10:16:25.407 18094 ERROR neutron.agent.dhcp_agent [-] Error importing interface driver 'None': 'NoneType' object has no attribute 'rpartition'
2015-01-23 10:16:25.613 18102 INFO neutron.common.config [-] Logging enabled!
2015-01-23 10:16:25.614 18102 ERROR neutron.agent.dhcp_agent [-] Error importing interface driver 'None': 'NoneType' object has no attribute 'rpartition'
2015-01-23 10:16:25.824 18110 INFO neutron.common.config [-] Logging enabled!
2015-01-23 10:16:25.825 18110 ERROR neutron.agent.dhcp_agent [-] Error importing interface driver 'None': 'NoneType' object has no attribute 'rpartition'
2015-01-23 10:16:26.032 18118 INFO neutron.common.config [-] Logging enabled!
2015-01-23 10:16:26.033 18118 ERROR neutron.agent.dhcp_agent [-] Error importing interface driver 'None': 'NoneType' object has no attribute 'rpartition'
2015-01-23 10:16:26.244 18126 INFO neutron.common.config [-] Logging enabled!
...

We believe the problem here is similar to the one described for Felix
in https://github.com/Metaswitch/calico/issues/137.
I.e. neutron-dhcp-agent is being rapidly respawned, and Chef's logic
for checking whether the service is running, and for restarting it,
interacts badly with that respawning.

A plausible fix for this is simply to delay the restart until later,
because, by that time, neutron-dhcp-agent will have already been
respawned and so restarted with a new set of config that allows it to
keep running - hence ending the rapid respawn cycle.  At that point,
therefore, Chef's restart logic should work.

But in that case why do we still need the restart at all, I hear you
ask?  The answer is that there might be an intermediate config state
that is picked up by neutron-dhcp-agent and allows it to keep running,
but which is still not the _correct_ config for Calico operation.
Therefore we do still want an explicit neutron-dhcp-agent restart, at
the end of the Chef provisioning.